Added Doors

+++ Vector2D.cpp
@@ -1,7 +1,7 @@
 #include "Vector2D.h"
 
-Vector2D::Vector2D(): x(0.0f), y(0.0f) {}
-Vector2D::Vector2D(float x, float y): x(x), y(y) {}
+Vector2D::Vector2D(): x(0), y(0) {}
+Vector2D::Vector2D(int x, int y): x(x), y(y) {}
 
 std::ostream& operator<<(std::ostream& stream, const Vector2D& vec)
 {
